Rename target for libmagic to avoid liblib
By naming the target "libmagic" we ended up with "liblibmagic.so". This
is because cmake prepends the library prefix to the target name for the
installation. We could tell cmake not to prepend the library name but
various stackoverflow users report mixed success and other weirdness
that can cause. A simpler solution is to just rename the target.